修改价格为bigDecimal-秒杀、拼团

+++ fightGroupModule/src/main/java/com/github/chenlijia1111/fightGroup/common/requestVo/fightGroupProduct/FightGroupProductAddParams.java
@@ -8,6 +8,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -54,7 +55,7 @@ public class FightGroupProductAddParams {
      */
     @ApiModelProperty("拼团价格")
     @PropertyCheck(name = "拼团价格", checkFunction = PriceCheck.class)
-    private Double fightPrice;
+    private BigDecimal fightPrice;
 
     /**
      * 拼团商品参与总库存
