+++ spikeModule/src/main/java/com/github/chenlijia1111/spike/common/response/product/SpikeAdminProductVo.java
@@ -6,6 +6,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -38,6 +39,6 @@ public class SpikeAdminProductVo extends AdminProductVo {
      */
     @ApiModelProperty("秒杀价格")
     @PropertyCheck(name = "秒杀价格")
-    private Double spikePrice;
+    private BigDecimal spikePrice;
 
 }
